Observer Updates and Game State
Map now populates correctly after selecting a game.
The player waiting view shows which players are currently joined in the
game, it does not update after the initial call to open the player
waiting window.  AI is listed in the window but no opportunity to set
an AI in the game.

I also added a GameState enum.