Undid Proxy Change so it can check for model
Instead of changing the proxy server to always update to get the
initial game state, change our start game state to -1.  This way when
we do the first game get, the server will have a version of 0 and we
will have a version of -1 ==> -1 < 0 is true so it will send us the
model.