State updates according to poller
Updated the states to update according to the state given in the model
every time the poller gets a new one.  Player waiting works, the views
update with each newly added player and the player waiting closes once
four players have joined and the state changes to the state listed in
the model.